chat: break out assistant message, clean up metadata tooltips
